feat: Initial creationg of logs engine
- Creation of logs engine as placeholder now
- Upgrading to zig 0.14.0 and fixing all the dependency problems
- Adding more checks to configs
- Adding "local" agent name to point master server it-self
- Adding compression types for rotate strategy
- Adding null checks because now some config values can be nullable

+++ src/app/tase.zig
@@ -37,16 +37,18 @@ pub const Tase = struct {
     }
 
     pub fn run(self: Tase) !void {
-        // try self.performCheck();
+        try self.performCheck();
         helpers.printApplicationInfo(self.version);
         if (self.cli_args.agent)
             try self.server.startAgentServer()
         else {
-            self.server.setAgents(self.yaml_cfg.agents);
-            // try self.server.startMasterServer();
-            const client = try clientFactory.getClient(self.allocator, "tcp", "127.0.0.1", 7423, "b9d36fa4b6cd3d8a2f5527c792143bfc");
-            defer client.destroy(self.allocator);
-            try client.sendMessage(&self.yaml_cfg.*.configs[0], self.allocator);
+            if (self.yaml_cfg.agents != null) {
+                self.server.setAgents(self.yaml_cfg.agents.?);
+                // try self.server.startMasterServer();
+                const client = try clientFactory.getClient(self.allocator, "tcp", "127.0.0.1", 7423, "b9d36fa4b6cd3d8a2f5527c792143bfc");
+                defer client.destroy(self.allocator);
+                try client.sendMessage(&self.yaml_cfg.*.configs[0], self.allocator);
+            }
         }
     }
